Load hypothesis client after ingestPdf is loaded
