+++ features/ati/AtiManuscript/IngestPdf/index.tsx
@@ -1,6 +1,7 @@
 import { FC, useState } from "react"
 
 import { InlineNotification } from "carbon-components-react"
+import DOMPurify from "isomorphic-dompurify"
 
 import { Document, Page, pdfjs } from "react-pdf"
 import { SizeMe } from "react-sizeme"
@@ -16,6 +17,17 @@ const IngestPdf: FC<IngestPdfProps> = ({ pdf }) => {
   const [numPages, setNumPages] = useState<number>(0)
 
   const onDocumentLoadSuccess = ({ numPages }: { numPages: number }) => {
+    // load hypothesis client
+    const hClientConfig = document.createElement("script")
+    hClientConfig.type = "application/json"
+    hClientConfig.className = "js-hypothesis-config"
+    hClientConfig.innerHTML = DOMPurify.sanitize(JSON.stringify({ openSidebar: true }))
+    const hClient = document.createElement("script")
+    hClient.src = "https://hypothes.is/embed.js"
+    hClient.async = true
+    document.head.appendChild(hClientConfig)
+    document.head.appendChild(hClient)
+
     setNumPages(numPages)
   }
